+++ target_postgres/connector.py
@@ -1,14 +1,10 @@
 """Connector class for target."""
 from __future__ import annotations
 
-from os import chmod, path
-from typing import cast
-
 import sqlalchemy
 from singer_sdk import SQLConnector
 from singer_sdk import typing as th
 from sqlalchemy.dialects.postgresql import ARRAY, BIGINT, JSONB
-from sqlalchemy.engine import URL
 from sqlalchemy.types import TIMESTAMP
 
 
@@ -70,27 +66,6 @@ class PostgresConnector(SQLConnector):
             A newly created SQLAlchemy engine object.
         """
         return self.create_sqlalchemy_engine().connect()
-
-    def get_sqlalchemy_url(self, config: dict) -> str:
-        """Generate a SQLAlchemy URL.
-
-        Args:
-            config: The configuration for the connector.
-        """
-        if config.get("sqlalchemy_url"):
-            return cast(str, config["sqlalchemy_url"])
-
-        else:
-            sqlalchemy_url = URL.create(
-                drivername=config["dialect+driver"],
-                username=config["user"],
-                password=config["password"],
-                host=config["host"],
-                port=config["port"],
-                database=config["database"],
-                query=self.get_sqlalchemy_query(config),
-            )
-            return cast(str, sqlalchemy_url)
 
     def drop_table(self, table: sqlalchemy.Table):
         """Drop table data."""
@@ -224,70 +199,3 @@ class PostgresConnector(SQLConnector):
                 "column_type": column.type.compile(dialect=self._engine.dialect),
             },
         )
-
-    def get_sqlalchemy_query(self, config: dict) -> dict:
-        """Get query values to be used for sqlalchemy URL creation.
-
-        Args:
-            config: The configuration for the connector.
-
-        Returns:
-            A dictionary with key-value pairs for the sqlalchemy query.
-        """
-        query = {}
-
-        # ssl_enable is for verifying the server's identity to the client.
-        if config["ssl_enable"]:
-            ssl_mode = config["ssl_mode"]
-            query.update({"sslmode": ssl_mode})
-            query["sslrootcert"] = self.filepath_or_certificate(
-                value=config["ssl_certificate_authority"],
-                alternative_name=config["ssl_storage_directory"] + "/root.crt",
-            )
-
-        # ssl_client_certificate_enable is for verifying the client's identity to the
-        # server.
-        if config["ssl_client_certificate_enable"]:
-            query["sslcert"] = self.filepath_or_certificate(
-                value=config["ssl_client_certificate"],
-                alternative_name=config["ssl_storage_directory"] + "/cert.crt",
-            )
-            query["sslkey"] = self.filepath_or_certificate(
-                value=config["ssl_client_private_key"],
-                alternative_name=config["ssl_storage_directory"] + "/pkey.key",
-                restrict_permissions=True,
-            )
-        return query
-
-    def filepath_or_certificate(
-        self,
-        value: str,
-        alternative_name: str,
-        restrict_permissions: bool = False,
-    ) -> str:
-        """Provide the appropriate key-value pair based on a filepath or raw value.
-
-        For SSL configuration options, support is provided for either raw values in
-        .env file or filepaths to a file containing a certificate. This function
-        attempts to parse a value as a filepath, and if no file is found, assumes the
-        value is a certificate and creates a file named `alternative_name` to store the
-        file.
-
-        Args:
-            value: Either a filepath or a raw value to be written to a file.
-            alternative_name: The filename to use in case `value` is not a filepath.
-            restrict_permissions: Whether to restrict permissions on a newly created
-                file. On UNIX systems, private keys cannot have public access.
-
-        Returns:
-            A dictionary with key-value pairs for the sqlalchemy query
-
-        """
-        if path.isfile(value):
-            return value
-        else:
-            with open(alternative_name, "wb") as alternative_file:
-                alternative_file.write(value.encode("utf-8"))
-            if restrict_permissions:
-                chmod(alternative_name, 0o600)
-            return alternative_name
